Prevent NoneType() >= int()
I encountered an edge case where the scanned domain used 'maxage' instead of 'max-age' in its HSTS header. This results in pshtt listing 'True' in the `HSTS` column but having an empty `HSTS Max Age` column. This resulted in an `TypeError: unorderable types: NoneType() >= int()` error caused by line 581. We can prevent this by checking if `hsts_age` is `True` in the in the `if`-statement leading up to that line.

This has the added advantage that HSTS headers with a `max-age` of 0 are also listed HSTS value 0 (no HSTS) instead of HSTS value 1 (HSTS with too low max-age). The [RFC specifies](https://tools.ietf.org/html/rfc6797#section-6.1.1) that a `max-age` of 0 means that the HSTS policy should be removed by the user agent. So the intended effect of the website is to have no HSTS. In a scan of 23k domains I actually got 44 domains listing a `max-age` of 0.