Add third-party and a11y to normal scan
I reverted the previous PR because I forgot the a11y config files. These
two scans take ~45 minutes combined, so this _shouldn't_ impact things
too much. Once we've confirmed that the data looks good, we'll switch
back to the normal scan and only run a11y and customer satisfaction on
the first of the month.